Feedback 2 revisions: 1) renamed the class name 2) added underscore for private fields 3) updated 'ThenNoError' for test methods 4) used var 5) added a small summary for all the test methods
